requirements.txt
# Backend Dependencies

# FastAPI and Core Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
mysqlclient>=2.2.1
aiomysql>=0.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.26.0

# Development
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Production
gunicorn>=21.2.0
uvicorn[standard]>=0.27.0 